Fixed 32/64-bit error. Fixed checking of array parent for subtable roots. A few calls to Verify() removed because they had to be.

+++ src/ArrayBinary.cpp
@@ -88,7 +88,7 @@ void ArrayBinary::Delete(size_t ndx) {
 
 	m_blob.Delete(start, end);
 	m_offsets.Delete(ndx);
-	m_offsets.Adjust(ndx, start - end);
+	m_offsets.Adjust(ndx, int64_t(start) - end);
 }
 
 void ArrayBinary::Resize(size_t ndx) {
